+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file. For commit 
  **Notes**:
  - desktop context menu "select multiple" enabled as optional user default (#1000)
  - onlyoffice readonly document types (".pages", ".numbers", ".key") list (#1018)
+ - onlyoffice tweaks to make more consistent, added logging (#1015)
 
  **BugFixes**:
  - fix lightBackground issue (#1021)